H8a {SiVHMP1}	R Documentation

Tablica H8a Medunarodne pricuve i devizna likvidnost Sluzbene medunarodne pricuve i ostale devizne pricuve (priblizna trzisna vrijednost)

Description

Tablica H8a Medunarodne pricuve i devizna likvidnost Sluzbene medunarodne pricuve i ostale devizne pricuve (priblizna trzisna vrijednost)

Usage


H8a

Format

An object of class data.frame with 228 rows and 19 columns.

Author(s)

David Ravnik dravnik@unipu.hr

Tomislava Kolar tkolar@unipu.hr

References

https://www.hnb.hr/statistika

[Package SiVHMP1 version 0.1.0 Index]